requirements.txt
# Database
pyodbc==5.0.1
pymssql==2.2.11

# Vector Embeddings
openai==1.12.0

# API Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Data Processing
pandas==2.2.0
numpy==1.26.4

# Environment Management
python-dotenv==1.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Utilities
tenacity==8.2.3
loguru==0.7.2
